Add more test coverage plus options in the sqlite package.

+++ sqlite/db.go
@@ -5,6 +5,7 @@ import (
 	"database/sql"
 	"time"
 
+	"github.com/benbjohnson/clock"
 	_ "github.com/mattn/go-sqlite3"
 	"github.com/pkg/errors"
 
@@ -13,8 +14,10 @@ import (
 
 // DB is an implementation of fab.HashDB that uses a Sqlite3 file for persistent storage.
 type DB struct {
-	db   *sql.DB
-	keep time.Duration
+	db             *sql.DB
+	keep           time.Duration
+	clk            clock.Clock
+	updateOnAccess bool
 }
 
 var _ fab.HashDB = &DB{}
@@ -24,6 +27,8 @@ CREATE TABLE IF NOT EXISTS hashes (
   hash BLOB NOT NULL PRIMARY KEY,
   unix_secs INT NOT NULL
 );
+
+CREATE INDEX IF NOT EXISTS unix_secs_idx ON hashes (unix_secs);
 `
 
 // Open opens the given file and returns it as a *DB.
@@ -39,9 +44,15 @@ func Open(ctx context.Context, path string, opts ...Option) (*DB, error) {
 	if err != nil {
 		return nil, errors.Wrap(err, "creating schema")
 	}
-	result := &DB{db: db}
+	result := &DB{
+		db:             db,
+		updateOnAccess: true,
+	}
 	for _, opt := range opts {
 		opt(result)
+	}
+	if result.clk == nil {
+		result.clk = clock.New()
 	}
 	return result, nil
 }
@@ -63,16 +74,45 @@ func Keep(d time.Duration) Option {
 	}
 }
 
+// WithClock is an Option that sets the database's clock.
+// By default it's clock.New(),
+// i.e. the normal time-telling clock.
+// For testing this can be set to a mock clock.
+func WithClock(clk clock.Clock) Option {
+	return func(db *DB) {
+		db.clk = clk
+	}
+}
+
+// UpdateOnAccess is an Option controlling whether to update a db entry's timestamp when accessed with Has.
+// The default is true: each Has of a value refreshes its timestamp to prevent its expiration.
+func UpdateOnAccess(update bool) Option {
+	return func(db *DB) {
+		db.updateOnAccess = update
+	}
+}
+
 // Has tells whether db contains the given hash.
 // If found, it also updates the last-access time of the hash.
 func (db *DB) Has(ctx context.Context, h []byte) (bool, error) {
-	const q = `UPDATE hashes SET unix_secs = $1 WHERE hash = $2`
-	res, err := db.db.ExecContext(ctx, q, time.Now().Unix(), h)
-	if err != nil {
-		return false, errors.Wrap(err, "updating database")
+	if db.updateOnAccess {
+		const q = `UPDATE hashes SET unix_secs = $1 WHERE hash = $2`
+		now := db.clk.Now()
+		res, err := db.db.ExecContext(ctx, q, now.Unix(), h)
+		if err != nil {
+			return false, errors.Wrap(err, "updating database")
+		}
+		aff, err := res.RowsAffected()
+		return aff > 0, errors.Wrap(err, "counting affected rows")
 	}
-	aff, err := res.RowsAffected()
-	return aff > 0, errors.Wrap(err, "counting affected rows")
+
+	const q = `SELECT COUNT(*) FROM hashes WHERE hash = $1`
+	var count int
+	err := db.db.QueryRowContext(ctx, q, h).Scan(&count)
+	if err != nil {
+		return false, errors.Wrap(err, "querying database")
+	}
+	return count > 0, nil
 }
 
 // Add adds a hash to db.
@@ -81,13 +121,15 @@ func (db *DB) Has(ctx context.Context, h []byte) (bool, error) {
 // entries with old last-access times are evicted.
 func (db *DB) Add(ctx context.Context, h []byte) error {
 	const q = `INSERT INTO hashes (hash, unix_secs) VALUES ($1, $2) ON CONFLICT DO UPDATE SET unix_secs = $2 WHERE hash = $1`
-	_, err := db.db.ExecContext(ctx, q, h, time.Now().Unix())
+	now := db.clk.Now()
+	_, err := db.db.ExecContext(ctx, q, h, now.Unix())
 	if err != nil {
 		return errors.Wrap(err, "adding hash to database")
 	}
 	if db.keep > 0 {
 		const q2 = `DELETE FROM hashes WHERE unix_secs < $1`
-		_, err = db.db.ExecContext(ctx, q2, time.Now().Add(-db.keep).Unix())
+		when := now.Add(-db.keep).Unix()
+		_, err = db.db.ExecContext(ctx, q2, when)
 		if err != nil {
 			return errors.Wrap(err, "evicting expired database entries")
 		}
